+++ src/draw.h
@@ -13,10 +13,10 @@ SDL_Surface* const_png_to_surface(const void* mem, int size);
     Returns: A newly allocated SDL_Texture, owned by the caller, or NULL if there was an error.
     Check the SDL error with the ERROR() macro.
  */
-SDL_Texture* const_png_to_texture(SDL_Renderer* renderer, const void* mem, int size);
+SDL_Texture* const_png_to_texture(SDL_Renderer* renderer, const void* mem, size_t size);
 
 /*
     Loads a texture from const memory and adds it to the state for later reference.
     Returns: 0 if successful
  */
-int texture_load_const_png(State* state, TexID texture_id, const void* mem, int size);
+int texture_load_const_png(State* state, TexID texture_id, const void* mem, size_t size);
